+++ src/shared/ArtistCalendar/AdditionalCalendarButtons.jsx
@@ -1,28 +1,11 @@
 import CustomButton from "../../package/customButton/CustomButton";
-import axiosInstance from "../services/axiosInstance";
+import { axiosDownloadFile } from "../services/axiosInstance";
 import useLinkUpStore from "../store/store";
 
 const downloadFile = async (artist_id) => {
     const queryParams = artist_id ? `?artist_id=${artist_id}` : "";
-    try {
-        const response = await axiosInstance.get(`/api/events/file/download-all${queryParams}`, {
-            responseType: "blob",
-        });
-
-        const downloadUrl = window.URL.createObjectURL(response.data);
-        const link = document.createElement("a");
-        link.href = downloadUrl;
-        link.download = "";
-
-        document.body.appendChild(link);
-        link.click();
-
-        document.body.removeChild(link);
-        window.URL.revokeObjectURL(downloadUrl);
-    } catch (error) {
-        console.error(error);
-        debugger;
-    }
+    const url = `/api/events/file/download-all${queryParams}`;
+    axiosDownloadFile(url);
 };
 
 export const BulkDownloadButton = () => {
