+++ src/shared/services/axiosInstance.js
@@ -53,3 +53,23 @@ export const axiosReturnsData = async (method, url, body, access_token) => {
             throw new Error("---- ERROR OCCURRED: 잘못된 메소드입니다");
     }
 };
+
+export const axiosDownloadFile = async (url) => {
+    try {
+        const response = await axiosInstance.get(url, { responseType: "blob" });
+
+        const downloadUrl = window.URL.createObjectURL(response.data);
+        const link = document.createElement("a");
+        link.href = downloadUrl;
+        link.download = "";
+
+        document.body.appendChild(link);
+        link.click();
+
+        document.body.removeChild(link);
+        window.URL.revokeObjectURL(downloadUrl);
+    } catch (error) {
+        console.error(error);
+        debugger;
+    }
+};
